fix(sentry): keep project-core under the file-size cap; fix the brief prompt
Two follow-ups on the #1748 brief redesign (#1769).

1. File-size hard cap. Merging main (#1775 added ~74 lines) pushed
   scripts/sentry-triage-project-core.mjs to 1031 lines on the PR-merge
   commit, tripping the brief suite's 1,000-line cap test (a gap the local
   gate missed pre-merge). Extract the projected-issue rendering — owning-repo
   body/title, the alias comment, and the idempotency marker/back-link
   matchers — into scripts/sentry-triage-projection.mjs, and move the shared
   SHORT-ID validators (isValidShortId, sanitizeDuplicateIds) down to the
   text layer. project-core drops to ~816 lines.

   The projection module is imported directly by the projection leg
   (sentry-triage-project.mjs), NOT re-exported through the verdict contract,
   so it stays OUT of the untrusted triage agent's write-wrapper runtime
   closure (pinned by sentry-triage-agent-comment.test.mjs). No cycle: the
   renderer sits on the text layer and project-core does not import it.

2. Stale prompt/CI doc (Codex #1769 round 4). The live triage prompt and the
   ci.yml brief-suite comment still said the brief is rendered into the issue
   BODY; after the redesign the leg only creates/updates a dedicated comment.
   Corrected both to describe the updated-in-place comment and reaffirm the
   single-body-writer invariant (archive owns the body; the brief owns its
   comment).
