+++ .github/prompts/sentry-triage.md
@@ -40,6 +40,6 @@ For a `needs-human` verdict, the escalation must be DECISION-READY — a human r
 - `investigated:` — a yaml list of what you already checked/ruled out (payload evidence read, code paths inspected, duplicate search run).
 - `escalation_reason:` — why you could not reach a confident verdict (ambiguity / security-sensitive surface / conflicting evidence).
 
-Keep each of those fields to ONE point, and do not restate them in the prose paragraph below the yaml block. A deterministic step renders these fields into the queue issue body as the human's brief, so a fifth paragraph repeating the same situation only pushes the decision further down the page. For a `needs-human` verdict the prose after the yaml block is at most two sentences of context the fields above do not already carry, or nothing at all.
+Keep each of those fields to ONE point, and do not restate them in the prose paragraph below the yaml block. A deterministic step renders these fields as a dedicated, updated-in-place COMMENT on the queue issue — the human's brief. (It never writes the issue body: the archive leg is the single body writer, and the brief owns its own comment instead.) A fifth paragraph repeating the same situation only pushes the decision further down the page, so for a `needs-human` verdict the prose after the yaml block is at most two sentences of context the fields above do not already carry, or nothing at all.
 
 Hard rules: max ~10 Sentry MCP calls; no Sentry mutations; no pushes/PRs/file edits; if anything fails irrecoverably, post a needs-human verdict explaining the failure rather than exiting silently.
