+++ .github/workflows/ci.yml
@@ -1068,12 +1068,14 @@ jobs:
         # the daily "default workflow-token permission stays read-only" check.
         run: node scripts/check-workflow-permissions-drift.test.mjs
       - name: Sentry needs-human brief suite
-        # Behavioral coverage of scripts/sentry-triage-brief.mjs (issue #1748):
-        # the rendered brief's fixed order and neutralization, the idempotent
-        # body write, and — the part only a test can hold — that every stub-body
-        # parser still reads the same values after the brief lands above the
-        # metadata yaml. It also asserts the prompt, the pipeline note and the
-        # triage workflow still agree with the contract it renders.
+        # Behavioral coverage of scripts/sentry-triage-brief.mjs (issue #1748,
+        # redesigned #1769): the rendered brief's fixed order and neutralization,
+        # the updated-in-place COMMENT lifecycle (created on needs-human, updated
+        # on re-triage, deleted on any other verdict), and — the part only a test
+        # can hold — that the leg NEVER writes the stub body, so it cannot drop
+        # the archive freshness baseline in the archive's unlabeled settlement
+        # window. It also asserts the prompt, the pipeline note and the triage
+        # workflow still agree with the contract it renders.
         run: pnpm sentry:brief:test
       - name: Override prune report self-tests
         # Behavioral coverage of scripts/override-prune-report.mjs (verdict
